Duke A-MoreOOP : Use More OOP
