+++ backup.txt
@@ -1,4 +1,3 @@
-E : 1 : exhibit : 2.30pm
 D : 1 : return book : June 6th
 D : 0 : attend exam : Fri noon
 E : 1 : attend a movie : 2.30pm HUB
