+++ src/main/java/ip/duke/Todo.java
@@ -9,7 +9,7 @@ import ip.duke.task.Task;
  * @see Task
  */
 public class Todo extends Task {
-    protected Todo(String description) {
+    public Todo(String description) {
         super(description);
         setId();
     }
