+++ src/main/java/ip/duke/storage/Wrapper.java
@@ -1,0 +1,5 @@
+package ip.duke.storage;
+
+public interface Wrapper {
+    void load();
+}
