+++ src/main/java/ip/duke/tasklist/TaskList.java
@@ -1,0 +1,190 @@
+package ip.duke.tasklist;
+
+import ip.duke.Deadline;
+import ip.duke.Event;
+import ip.duke.Todo;
+import ip.duke.exceptions.DukeException;
+import ip.duke.storage.Storage;
+import ip.duke.task.Task;
+import ip.duke.ui.Ui;
+
+import java.io.*;
+import java.util.LinkedHashSet;
+
+public class TaskList {
+    // Collection preserves input sequence, yields constant time for update/delete ops
+    private final static LinkedHashSet<Task> TASKS = new LinkedHashSet<>(125, (float) 0.8);
+
+    public static int listSize(){
+        return TASKS.size();
+    }
+
+    public static LinkedHashSet<Task> getList(){
+        return TASKS;
+    }
+
+    //for possible closure technical glitch
+    public static void toClose(Closeable obj) {
+
+        if (obj != null) {
+
+            try {
+                obj.close();
+
+            } catch (IOException ex) {
+                System.out.print("LisGenie : ");
+                System.out.println("Possible disc error / file system full!" + ex.getMessage());
+            }
+        }
+    }
+    /**
+     * Extract a Task item by "index" or its sequenced order.
+     * @param idx indexed location of target Task item
+     * @return A Task object.
+     */
+    public static Task getItem(int idx){
+        int currentIndex = 0;
+        Task item = null;
+
+        for(Task element : TASKS){
+            if(currentIndex == idx){
+                item = element;
+                break;
+            }else{
+                ++currentIndex;
+            }
+        }
+        return item;
+    }
+
+    public static void doDelete(String word) {
+        int itemIndex = Integer.parseInt(word.trim()) - 1;
+
+        if (itemIndex < 0 || itemIndex > 99) {
+            Ui.echoOffList(itemIndex);
+        } else {
+            drop(itemIndex);
+            Storage.writeToFile();
+        }
+    }
+
+    public static void doDone(String word) {
+        int itemIndex = Integer.parseInt(word.trim()) - 1;
+
+        if (itemIndex < 0 || itemIndex > 99) {
+            Ui.echoOffList(itemIndex);
+        } else {
+            updateDoneStatus(itemIndex);
+            Storage.writeToFile();
+        }
+    }
+
+    public static void drop(int idx){
+        Task item = getItem(idx);
+
+        if(item != null) {
+            TASKS.remove(item);
+            Ui.echoDelete(item);
+        }else{
+            Ui.echoNoEntries();
+        }
+    }
+
+    private static Task createTask(String str) {
+        String[] text = str.trim().split(":");
+
+        for (int i = 0; i < text.length; i++) {
+            text[i] = text[i].trim();
+        }
+
+        Task t = null;
+
+        try {
+            switch (text[0]) {
+            case "T":
+                t = new Todo(text[2]);
+                break;
+            case "D":
+                t = new Deadline(text[2], text[3]);
+                break;
+            case "E":
+                t = new Event(text[2], text[3]);
+                break;
+            default:
+                throw new IllegalStateException("Unexpected value: " + text[0]);
+            }
+
+            if (text[1].equals("1")) {
+                t.setDone();
+            }
+
+        } catch (ArrayIndexOutOfBoundsException | IllegalStateException err) {
+            System.out.println("Database file format errors detected: Contact Admin.");
+        }
+        return t;
+    }
+
+    public static LinkedHashSet<Task> load(String path) throws DukeException {  //load from file
+        BufferedReader reader = null;
+
+        try {
+            reader = new BufferedReader(new FileReader(path));
+            String line = reader.readLine();
+
+            while (line != null && !line.trim().isEmpty()) {
+                Task entry = createTask(line.replace("\\s+", " "));
+
+                if (entry != null) {
+                    TaskList.TASKS.add(entry);
+                }
+
+                line = reader.readLine();
+            }
+
+        } catch (IOException e) {
+            throw new DukeException("File access/read error!", e);
+        } finally{
+            toClose(reader);
+        }
+        return TASKS;
+    }
+
+    public static void updateDoneStatus(int idx){
+        Task item = getItem(idx);
+
+        if(item != null) {
+            if (!item.getStatusIcon().equals(" ")) {
+                System.out.print("LisGenie : ");
+                System.out.println("Mmm...item already marked done, O Master? Next @|@ task?");
+            } else {
+                item.setDone();
+                Ui.echoDone(item);
+            }
+        }else{
+            Ui.echoNoEntries();
+        }
+    }
+
+    public static void addDeadline(String input) {
+        String[] parts = input.split(" /by ", 2);
+        Deadline item = new Deadline(parts[0].trim(), parts[1].trim());
+        TASKS.add(item);
+        Ui.echoAdded(item);
+        Storage.appendToFile(TASKS.size() - 1);
+    }
+
+    public static void addEvent(String input) {
+        String[] parts = input.split(" /at ", 2);
+        Event item = new Event(parts[0].trim(), parts[1].trim());
+        TASKS.add(item);
+        Ui.echoAdded(item);
+        Storage.appendToFile(TASKS.size() - 1);
+    }
+
+    public static void addTodo(String input) {
+        Todo item = new Todo(input);
+        TASKS.add(item);
+        Ui.echoAdded(item);
+        Storage.appendToFile(TASKS.size() - 1);
+    }
+}
